tickets/src/routes/new.ts
import express, { Request, Response } from 'express';
import { requireAuth, validateRequest } from '@jingjingthinkingtickets/common';
import { body } from 'express-validator';
import { Ticket } from '../models/tickets';
const router = express.Router();

router.post(
  '/api/tickets',
  requireAuth,
  [
    body('title').not().isEmpty().withMessage('title is required'),
    body('price').isFloat({ gt: 0 }).withMessage('price is invalid'),
  ],
  validateRequest,
  async (req: Request, res: Response) => {
    const { title, price } = req.body;
    const ticket = Ticket.build({
      title,
      price,
      userId: req.currentUser!.id,
    });
    await ticket.save();
    res.status(201).send(ticket);
  }
);

export { router as createTicketRouter };
